[feat] Add spike bitwise operations and update module exports
